spark_hm_1/requirements.txt
attrs==21.2.0
azure-core==1.17.0
azure-storage-blob==12.9.0b1
azure-storage-file-datalake==12.4.0
backoff==1.11.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
idna==3.2
iniconfig==1.1.1
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
opencage==2.0.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
py4j==0.10.9
pycparser==2.20
pygeohash==1.2.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyspark==3.1.2
pytest==6.2.5
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
